refactor: reorganize NameState usage and move related structs to web module
